Add grace period for CLI daemon
This adds a 3 second grace period before playing alarms when running the
CLI as a daemon. This allows UIs like Aevum to remove the alarm while
starting playback itself.

By having both of these clients coexist, it's possible to have a CLI
daemon in the background 24/7 while the UI takes priority whenever it is
running.