Move Aevum to workspace root
Since Aevum is the primary focus of the repository, it seems appropriate
to have it as the workspace root member rather than the alarm dependency
shared by the other crates.

While no root member would be another option, this should make it
easiest to build what most people would be interested in.